test: create `test_remote` & add `test_delete_records`
